workflow: set default python version to 3.8
There is an issue with python 3.9.0 not working with the numpy version used in requirements.txt. I am testing whether this is true by changing the default python version used by the workflow to 3.8. It can't be tested on a separate branch, because workflows only work on the master branch.